Handle blank authors from mtn
Added blank entry to authors_map.txt

Also, modify check_authors so as not to ignore blank authors.

This way, mtn git_export won't fail due to a blank author name that
appears in the Pidgin repo.  This seems not to have been a problem with
earlier versions of mtn, but mtn v0.48 barfs if it encounters a blank
author without a corresponding entry in the author map.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>
Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>